+++ DocumentsApp/src/com/rembli/dms/File.java
@@ -3,7 +3,7 @@ import lombok.Data;
 
 @Data
 public class File {
-	protected int idfile;
+	protected long idfile;
 	protected String fileName;
 	protected String fileType;
 	protected byte[] data;
